refactor: 분석 플로우 책임 추가 분리
